backend/requirements.txt
asgiref==3.4.1
attrs==22.2.0
autowrapt==1.0
basictracer==3.2.0
bleach==4.1.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==35.0.0
Django==4.0.6
django-appconf==1.0.5
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
drf-api-logger==1.0.8
drf-spectacular==0.24.2
fysom==2.1.6
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
instana==1.38.0
jsonschema==4.17.3
opentracing==2.4.0
packaging==21.0
phonenumbers==8.12.33
protobuf==3.20.3
psycopg2-binary==2.9.1
pycparser==2.20
pyparsing==2.4.7
pyrsistent==0.19.3
pytz==2021.1
PyYAML==6.0
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
webencodings==0.5.1
wrapt==1.14.1